requirements.txt
langchain==1.0.5
langchain-core==1.0.4
langchain-community==0.4.1
langchain-huggingface==1.0.1
langchain-openai==1.0.2


chromadb==1.3.4
pymilvus==2.6.3


sentence-transformers==5.1.2
transformers==4.57.3
torch==2.9.1


pymongo==4.15.4


fastapi==0.121.2
uvicorn==0.38.0
python-dotenv==1.2.1
pydantic==2.12.4
pydantic-settings==2.12.0
python-multipart==0.0.20


pillow==9.5.0
pdfplumber==0.11.8
unstructured==0.10.10
unstructured[local-inference]==0.10.10


numpy==1.26.4
scikit-learn==1.7.2
matplotlib==3.10.7


mlflow==3.6.0


streamlit>=1.20.0
requests>=2.28.0